requirements.txt
appdirs==1.4.3
asgiref==3.2.7
azt==0.0.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Cython==0.29.16
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.7
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
importlib-metadata==1.6.0
kombu==4.6.8
lxml==4.5.0
multitasking==0.0.9
msgpack==0.6.2
numpy==1.18.2
pandas==1.0.3
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pinyinxy1==1.0.0
psycopg2-binary==2.8.4
pyfiglet==0.8.post1
pypinyin==0.37.0
python-dateutil==2.8.1
pytz==2020.1
pyzt==1.0.1
redis==3.5.0
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
zipp==3.1.0

djangorestframework~=3.11.0